requirements.txt
-i https://pypi.org/simple
apscheduler==3.10.4
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cryptography==41.0.5
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.1.5
django-allauth==0.51.0
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
environs==8.0.0
gunicorn==19.9.0
idna==3.4; python_version >= '3.5'
marshmallow==3.20.1; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pillow==9.2.0
psycopg2-binary==2.9.5
pycparser==2.21
pyjwt==2.8.0; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.8.0; python_version < '3.11'
tzlocal==5.2; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.7'
whitenoise==5.1.0
